Add web UI for viewing aggregated trace data
After running the zipkin/aggregate hadoop job, dependency
links are stored in the database. This allows us to draw a
graph of the entire system.

Hovering over nodes will highlight which dependency links it has.
Clicking a node will show a list of its dependency links, and
from there you can click on a dependency to view data about the
dependency link itself. (number of calls, mean time etc)

The width of an arrow corresponds to the amount of traffic
between nodes (the number of calls made).